requirements-production.txt
# TAAXDOG Production Requirements
# Extends base requirements.txt with production-specific dependencies

# Base requirements (inherit from requirements.txt)
-r requirements.txt

# Production Web Server
gunicorn==22.0.0
gevent==24.2.1

# Caching and Performance
redis==5.0.8
flask-caching==2.3.0
flask-compress==1.15

# Security and Rate Limiting
flask-limiter==3.8.0
flask-talisman==1.1.0
bcrypt==4.2.1
cryptography==45.0.3

# Monitoring and Error Tracking
sentry-sdk[flask]==2.17.0
newrelic==10.5.0
prometheus-flask-exporter==0.23.0

# Analytics and Metrics
mixpanel==5.0.0
segment-analytics-python==2.3.2

# Australian Business Integration
zeep==4.2.1  # For SOAP APIs (ATO integration)
requests-oauthlib==2.0.0

# Database Performance
psycopg2-binary==2.9.9  # PostgreSQL adapter (if using PostgreSQL)
sqlalchemy==2.0.35
alembic==1.14.0

# Background Tasks
celery==5.3.6
redis==5.0.8  # Redis as message broker

# Email and Notifications
flask-mail==0.10.0
sendgrid==6.11.0
twilio==9.3.9

# File Processing and Optimization
pillow==11.2.1
pdf2image==1.17.0
opencv-python==4.10.0.84

# Testing and Quality Assurance
pytest-cov==6.0.0
pytest-xdist==3.6.2
black==24.10.0
flake8==7.1.1
bandit==1.8.0  # Security linting

# Documentation
sphinx==8.1.3
sphinx-rtd-theme==3.0.2

# Development Tools (for staging)
memory-profiler==0.61.0
py-spy==0.3.14

# Notification system dependencies  
schedule==1.2.0 